made list component as a generic implementation for creating a list element for display solved problem when deserializing instances where event listeners would not deserialize
