+++ src/esportsapp/Event.java
@@ -14,10 +14,10 @@ import javax.swing.*;
 
 /**
  *
- * @author User
+ * @author j9neave
  */
-public class Event extends javax.swing.JPanel implements EsportsInterface<Event> {
-
+public class Event implements EsportsInterface<Event>{
+	
 	@Override
 	public Event get() {
 		return this;
@@ -31,7 +31,7 @@ public class Event extends javax.swing.JPanel implements EsportsInterface<Event>
 	@Override
 	public void setName(String name) {
 		this.name = name;
-		this.nameLabel.setText(name);
+		//this.nameLabel.setText(name);
 	}
 	
 	@Override
@@ -42,7 +42,7 @@ public class Event extends javax.swing.JPanel implements EsportsInterface<Event>
 	@Override 
 	public void setIcon(ImageIcon icn) {
 		this.icon = icn;
-		this.iconLabel.setIcon(new ImageIcon(this.icon.getImage().getScaledInstance(32, 32, Image.SCALE_DEFAULT)));
+		this.panel.getIconLabel().setIcon(new ImageIcon(this.icon.getImage().getScaledInstance(32, 32, Image.SCALE_DEFAULT)));
 	}
 	
 	@Override 
@@ -83,10 +83,8 @@ public class Event extends javax.swing.JPanel implements EsportsInterface<Event>
 	String description;
 	int score;
 	Map<Team, Integer> scores;
+	transient ListComponent<Event> panel;
 	
-	/**
-	 * Creates new form Event
-	 */
 	private Event(String name) {
 		this.name = name;
 		this.icon = Palette.getCurrentScheme().ICON_DEFAULT_EVENT.getIcon();
@@ -94,29 +92,21 @@ public class Event extends javax.swing.JPanel implements EsportsInterface<Event>
 		this.score = 0;
 		this.scores = new HashMap<>();
 		
-		
-		initComponents();
 	}
 	
 	public static Event add(String name) {
 		name = name.trim();
 		Event evt = new Event(name);
 		
+		// create panel and add to elp
+		evt.panel = new ListComponent<>(Event.class, evt);
 		JPanel elp = EsportsGUI.getEventsListPanel();
-		elp.add(evt, Event.list.size()); 
+		elp.add(evt.panel, elp.getComponentCount() - 1);
 		elp.revalidate();
 		
 		Event.list.add(evt);
 		
 		return evt;
-	}
-	
-	public JLabel getNameLabel() {
-		return nameLabel;
-	}
-	
-	public JLabel getIconLabel() {
-		return iconLabel;
 	}
 	
 	public void addTeam(Team tm) {
@@ -127,78 +117,5 @@ public class Event extends javax.swing.JPanel implements EsportsInterface<Event>
 		this.scores.put(tm, scr);
 		tm.updateScore();
 	}
-
-	/**
-	 * This method is called from within the constructor to initialize the form.
-	 * WARNING: Do NOT modify this code. The content of this method is always
-	 * regenerated by the Form Editor.
-	 */
-	@SuppressWarnings("unchecked")
-    // <editor-fold defaultstate="collapsed" desc="Generated Code">//GEN-BEGIN:initComponents
-    private void initComponents() {
-        java.awt.GridBagConstraints gridBagConstraints;
-
-        nameLabel = new javax.swing.JLabel();
-        iconLabel = new javax.swing.JLabel();
-
-        setBackground(Palette.getCurrentScheme().COLOR_ACTIVE.getColor());
-        setMaximumSize(new java.awt.Dimension(2147483647, 35));
-        setMinimumSize(new java.awt.Dimension(0, 35));
-        setPreferredSize(new java.awt.Dimension(400, 35));
-        addMouseListener(new java.awt.event.MouseAdapter() {
-            public void mouseClicked(java.awt.event.MouseEvent evt) {
-                formMouseClicked(evt);
-            }
-            public void mouseEntered(java.awt.event.MouseEvent evt) {
-                formMouseEntered(evt);
-            }
-            public void mouseExited(java.awt.event.MouseEvent evt) {
-                formMouseExited(evt);
-            }
-        });
-        setLayout(new java.awt.GridBagLayout());
-
-        nameLabel.setFont(new java.awt.Font("Tahoma", 1, 18)); // NOI18N
-        nameLabel.setForeground(Palette.getCurrentScheme().COLOR_PRIMARY_TEXT.getColor());
-        nameLabel.setText(this.name);
-        nameLabel.setBorder(javax.swing.BorderFactory.createEmptyBorder(5, 5, 5, 5));
-        gridBagConstraints = new java.awt.GridBagConstraints();
-        gridBagConstraints.gridx = 1;
-        gridBagConstraints.gridy = 0;
-        gridBagConstraints.fill = java.awt.GridBagConstraints.BOTH;
-        gridBagConstraints.anchor = java.awt.GridBagConstraints.WEST;
-        gridBagConstraints.weightx = 1.0;
-        gridBagConstraints.weighty = 1.0;
-        add(nameLabel, gridBagConstraints);
-
-        iconLabel.setIcon(new ImageIcon(this.icon.getImage().getScaledInstance(32, 32, Image.SCALE_DEFAULT)));
-        gridBagConstraints = new java.awt.GridBagConstraints();
-        gridBagConstraints.gridx = 0;
-        gridBagConstraints.gridy = 0;
-        gridBagConstraints.anchor = java.awt.GridBagConstraints.WEST;
-        add(iconLabel, gridBagConstraints);
-    }// </editor-fold>//GEN-END:initComponents
-
-    private void formMouseEntered(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_formMouseEntered
-        // TODO add your handling code here:
-		this.setBackground(Palette.getCurrentScheme().COLOR_MENU_MOUSE_OVER_BACK.getColor());
-    }//GEN-LAST:event_formMouseEntered
-
-    private void formMouseExited(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_formMouseExited
-        // TODO add your handling code here:
-		this.setBackground(Palette.getCurrentScheme().COLOR_ACTIVE.getColor());
-    }//GEN-LAST:event_formMouseExited
-
-    private void formMouseClicked(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_formMouseClicked
-        // TODO add your handling code here:
-		ManagementDialog<Event> md = new ManagementDialog<>(this);
-		md.setVisible(true);
-		md.nameField.setText(md.object.name);
-		
-    }//GEN-LAST:event_formMouseClicked
-
-    // Variables declaration - do not modify//GEN-BEGIN:variables
-    private javax.swing.JLabel iconLabel;
-    private javax.swing.JLabel nameLabel;
-    // End of variables declaration//GEN-END:variables
+	
 }
