+++ src/esportsapp/ListComponent.java
@@ -1,0 +1,117 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package esportsapp;
+
+import java.awt.*;
+import javax.swing.*;
+
+/**
+ *
+ * @author j9neave
+ * @param <T>
+ */
+public class ListComponent<T extends EsportsInterface> extends javax.swing.JPanel {
+	
+	Class<T> type;
+	T object;
+	
+	public ListComponent<T> get() {
+		return this;
+	}
+	
+	/**
+	 * Creates new form ListComponent
+	 * @param t
+	 * @param obj
+	 */
+	public ListComponent(Class<T> t, T obj) {
+		this.type = t;
+		this.object = obj;
+		
+		initComponents();
+	}
+	
+	public JLabel getIconLabel() {
+		return iconLabel;
+	}
+	
+	public JLabel getNameLabel() {
+		return nameLabel;
+	}
+
+	/**
+	 * This method is called from within the constructor to initialize the form.
+	 * WARNING: Do NOT modify this code. The content of this method is always
+	 * regenerated by the Form Editor.
+	 */
+	@SuppressWarnings("unchecked")
+    // <editor-fold defaultstate="collapsed" desc="Generated Code">//GEN-BEGIN:initComponents
+    private void initComponents() {
+        java.awt.GridBagConstraints gridBagConstraints;
+
+        iconLabel = new javax.swing.JLabel();
+        nameLabel = new javax.swing.JLabel();
+
+        setBackground(Palette.getCurrentScheme().COLOR_ACTIVE.getColor());
+        setMaximumSize(new java.awt.Dimension(2147483647, 35));
+        setMinimumSize(new java.awt.Dimension(200, 35));
+        setPreferredSize(new java.awt.Dimension(400, 35));
+        addMouseListener(new java.awt.event.MouseAdapter() {
+            public void mouseClicked(java.awt.event.MouseEvent evt) {
+                formMouseClicked(evt);
+            }
+            public void mouseEntered(java.awt.event.MouseEvent evt) {
+                formMouseEntered(evt);
+            }
+            public void mouseExited(java.awt.event.MouseEvent evt) {
+                formMouseExited(evt);
+            }
+        });
+        setLayout(new java.awt.GridBagLayout());
+
+        iconLabel.setIcon(new ImageIcon(this.object.getIcon().getImage().getScaledInstance(32, 32, Image.SCALE_DEFAULT)));
+        gridBagConstraints = new java.awt.GridBagConstraints();
+        gridBagConstraints.anchor = java.awt.GridBagConstraints.WEST;
+        add(iconLabel, gridBagConstraints);
+
+        nameLabel.setFont(new java.awt.Font("Tahoma", 1, 18)); // NOI18N
+        nameLabel.setForeground(Palette.getCurrentScheme().COLOR_PRIMARY_TEXT.getColor());
+        nameLabel.setText(this.object.getName());
+        nameLabel.setBorder(javax.swing.BorderFactory.createEmptyBorder(5, 5, 5, 5));
+        gridBagConstraints = new java.awt.GridBagConstraints();
+        gridBagConstraints.gridx = 1;
+        gridBagConstraints.gridy = 0;
+        gridBagConstraints.fill = java.awt.GridBagConstraints.BOTH;
+        gridBagConstraints.ipady = 3;
+        gridBagConstraints.anchor = java.awt.GridBagConstraints.WEST;
+        gridBagConstraints.weightx = 1.0;
+        gridBagConstraints.weighty = 1.0;
+        add(nameLabel, gridBagConstraints);
+    }// </editor-fold>//GEN-END:initComponents
+
+    private void formMouseEntered(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_formMouseEntered
+        // TODO add your handling code here:
+		this.setBackground(Palette.getCurrentScheme().COLOR_MENU_MOUSE_OVER_BACK.getColor());
+    }//GEN-LAST:event_formMouseEntered
+
+    private void formMouseExited(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_formMouseExited
+        // TODO add your handling code here:
+		this.setBackground(Palette.getCurrentScheme().COLOR_ACTIVE.getColor());
+    }//GEN-LAST:event_formMouseExited
+
+    private void formMouseClicked(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_formMouseClicked
+        // TODO add your handling code here:
+		ManagementDialog<T> md = new ManagementDialog<T>(this.object);
+		md.setVisible(true);
+		md.nameField.setText(md.object.getName());
+    }//GEN-LAST:event_formMouseClicked
+
+	
+    // Variables declaration - do not modify//GEN-BEGIN:variables
+    private javax.swing.JLabel iconLabel;
+    private javax.swing.JLabel nameLabel;
+    // End of variables declaration//GEN-END:variables
+}
