+++ src/esportsapp/Member.java
@@ -9,15 +9,14 @@ import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.awt.*;
-import javax.swing.*;
+import javax.swing.ImageIcon;
 
 /**
  *
- * @author User
+ * @author j9neave
  */
-public class Member extends javax.swing.JPanel implements EsportsInterface<Member> {
-
+public class Member implements EsportsInterface<Member> {
+	
 	@Override
 	public Member get() {
 		return this;
@@ -73,7 +72,7 @@ public class Member extends javax.swing.JPanel implements EsportsInterface<Membe
 		this.scores = mp;
 	}
 	
-	public static List<Member> list = new ArrayList<Member>();
+	public static List<Member> list = new ArrayList<>();
 	private static final long serialVersionUID = 46L;
 	
 	String name;
@@ -84,9 +83,6 @@ public class Member extends javax.swing.JPanel implements EsportsInterface<Membe
 	
 	Team team;
 	
-	/**
-	 * Creates new form Member
-	 */
 	private Member(String name, Team team) {
 		this.name = name;
 		this.icon = Palette.getCurrentScheme().ICON_DEFAULT_MEMBER.getIcon();
@@ -97,8 +93,6 @@ public class Member extends javax.swing.JPanel implements EsportsInterface<Membe
 		this.team = team;
 		this.score = 0;
 		
-		
-		initComponents();
 	}
 	
 	public static Member add(String name, Team team) {
@@ -120,46 +114,5 @@ public class Member extends javax.swing.JPanel implements EsportsInterface<Membe
 		this.team.scores.put(this, total);	
 		
 	}
-
-	/**
-	 * This method is called from within the constructor to initialize the form.
-	 * WARNING: Do NOT modify this code. The content of this method is always
-	 * regenerated by the Form Editor.
-	 */
-	@SuppressWarnings("unchecked")
-    // <editor-fold defaultstate="collapsed" desc="Generated Code">//GEN-BEGIN:initComponents
-    private void initComponents() {
-        java.awt.GridBagConstraints gridBagConstraints;
-
-        nameLabel = new javax.swing.JLabel();
-        iconLabel = new javax.swing.JLabel();
-
-        setPreferredSize(new java.awt.Dimension(400, 35));
-        setLayout(new java.awt.GridBagLayout());
-
-        nameLabel.setFont(new java.awt.Font("Tahoma", 1, 18)); // NOI18N
-        nameLabel.setForeground(Palette.getCurrentScheme().COLOR_PRIMARY_TEXT.getColor());
-        nameLabel.setText(this.name);
-        nameLabel.setBorder(javax.swing.BorderFactory.createEmptyBorder(5, 5, 5, 5));
-        gridBagConstraints = new java.awt.GridBagConstraints();
-        gridBagConstraints.gridx = 1;
-        gridBagConstraints.gridy = 0;
-        gridBagConstraints.anchor = java.awt.GridBagConstraints.NORTHWEST;
-        gridBagConstraints.weightx = 1.0;
-        gridBagConstraints.weighty = 1.0;
-        add(nameLabel, gridBagConstraints);
-
-        iconLabel.setIcon(new ImageIcon(this.icon.getImage().getScaledInstance(32, 32, Image.SCALE_DEFAULT)));
-        gridBagConstraints = new java.awt.GridBagConstraints();
-        gridBagConstraints.gridx = 0;
-        gridBagConstraints.gridy = 0;
-        gridBagConstraints.anchor = java.awt.GridBagConstraints.NORTHWEST;
-        add(iconLabel, gridBagConstraints);
-    }// </editor-fold>//GEN-END:initComponents
-
 	
-    // Variables declaration - do not modify//GEN-BEGIN:variables
-    private javax.swing.JLabel iconLabel;
-    private javax.swing.JLabel nameLabel;
-    // End of variables declaration//GEN-END:variables
 }
