+++ src/esportsapp/Team.java
@@ -14,9 +14,9 @@ import javax.swing.*;
 
 /**
  *
- * @author User
+ * @author j9neave
  */
-public class Team extends javax.swing.JPanel implements EsportsInterface<Team> {
+public class Team implements EsportsInterface<Team> {
 	
 	@Override
 	public Team get() {
@@ -81,10 +81,8 @@ public class Team extends javax.swing.JPanel implements EsportsInterface<Team> {
 	int score;
 	String description;
 	Map<Member, Integer> scores;
+	transient ListComponent<Team> panel;
 	
-	/**
-	 * Creates new form Team
-	 */
 	private Team(String name) {
 		this.name = name;
 		this.icon = Palette.getCurrentScheme().ICON_DEFAULT_TEAM.getIcon();
@@ -93,13 +91,16 @@ public class Team extends javax.swing.JPanel implements EsportsInterface<Team> {
 		
 		this.score = 0;
 		
-		
-		initComponents();
 	}
 	
 	public static Team add(String name) {
 		name = name.trim();
 		Team tm = new Team(name);
+		
+		tm.panel = new ListComponent<>(Team.class, tm);
+		JPanel tlp = EsportsGUI.getTeamsListPanel();
+		tlp.add(tm.panel, tlp.getComponentCount() - 1);
+		tlp.revalidate();
 		
 		Team.list.add(tm);
 		
@@ -114,53 +115,5 @@ public class Team extends javax.swing.JPanel implements EsportsInterface<Team> {
 				.reduce(tot, Integer::sum);
 		this.score = tot;
 	}
-
-	/**
-	 * This method is called from within the constructor to initialize the form.
-	 * WARNING: Do NOT modify this code. The content of this method is always
-	 * regenerated by the Form Editor.
-	 */
-	@SuppressWarnings("unchecked")
-    // <editor-fold defaultstate="collapsed" desc="Generated Code">//GEN-BEGIN:initComponents
-    private void initComponents() {
-        java.awt.GridBagConstraints gridBagConstraints;
-
-        nameLabel = new javax.swing.JLabel();
-        iconLabel = new javax.swing.JLabel();
-
-        setMaximumSize(new java.awt.Dimension(2147483647, 40));
-        setMinimumSize(new java.awt.Dimension(0, 40));
-        setOpaque(false);
-        setPreferredSize(new java.awt.Dimension(400, 35));
-        java.awt.GridBagLayout layout = new java.awt.GridBagLayout();
-        layout.columnWeights = new double[] {0.0};
-        layout.rowWeights = new double[] {0.0};
-        setLayout(layout);
-
-        nameLabel.setFont(new java.awt.Font("Tahoma", 1, 18)); // NOI18N
-        nameLabel.setForeground(Palette.getCurrentScheme().COLOR_PRIMARY_TEXT.getColor());
-        nameLabel.setText(this.name);
-        nameLabel.setBorder(javax.swing.BorderFactory.createEmptyBorder(5, 5, 5, 5));
-        gridBagConstraints = new java.awt.GridBagConstraints();
-        gridBagConstraints.gridx = 1;
-        gridBagConstraints.gridy = 0;
-        gridBagConstraints.fill = java.awt.GridBagConstraints.BOTH;
-        gridBagConstraints.anchor = java.awt.GridBagConstraints.WEST;
-        gridBagConstraints.weightx = 1.0;
-        gridBagConstraints.weighty = 1.0;
-        add(nameLabel, gridBagConstraints);
-
-        iconLabel.setIcon(new ImageIcon(this.icon.getImage().getScaledInstance(32, 32, Image.SCALE_DEFAULT)));
-        gridBagConstraints = new java.awt.GridBagConstraints();
-        gridBagConstraints.gridx = 0;
-        gridBagConstraints.gridy = 0;
-        gridBagConstraints.anchor = java.awt.GridBagConstraints.WEST;
-        add(iconLabel, gridBagConstraints);
-    }// </editor-fold>//GEN-END:initComponents
-
-
-    // Variables declaration - do not modify//GEN-BEGIN:variables
-    private javax.swing.JLabel iconLabel;
-    private javax.swing.JLabel nameLabel;
-    // End of variables declaration//GEN-END:variables
+	
 }
